from ptrace manual:
PTRACE_ATTACH
Attach  to  the  process specified in pid, making it a tracee of the calling process.  The tracee is sent a SIGSTOP, but will not necessarily have stopped by the completion of
this call; use waitpid(2) to wait for the tracee to stop.  See the "Attaching and detaching" subsection for additional information.  (addr and data are ignored.)